add IonIdentitiyMolecularNetworking class with annotateConsensusMap method
