+++ src/openms/source/ANALYSIS/ID/sources.cmake
@@ -30,6 +30,7 @@ IDRipper.cpp
 IDDecoyProbability.cpp
 IDScoreGetterSetter.cpp
 IDScoreSwitcherAlgorithm.cpp
+IonIdentityMolecularNetworking
 MessagePasserFactory.cpp
 MetaboliteSpectralMatching.cpp
 PeptideProteinResolution.cpp
