+++ src/topp/GNPSExport.cpp
@@ -35,6 +35,7 @@
 
 #include <OpenMS/APPLICATIONS/TOPPBase.h>
 #include <OpenMS/FORMAT/GNPSMGFFile.h>
+#include <OpenMS/ANALYSIS/ID/IonIdentityMolecularNetworking.h>
 
 using namespace OpenMS;
 using namespace std;
@@ -135,6 +136,8 @@ protected:
     gnps.setParameters(getParam_()); // copy tool parameter to library class/algorithm
     gnps.run(consensus_file_path, mzml_file_paths, out);
 
+    IonIdentityMolecularNetworking iimn;
+    iimn.logString("banana");
     return EXECUTION_OK;
   }
 };
